+++ rpc/namespaces/ethereum/eth/api.go
@@ -405,12 +405,23 @@ func (e *PublicAPI) GetTransactionLogs(txHash common.Hash) ([]*ethtypes.Log, err
 		return nil, nil
 	}
 
-	msgIndex, _ := rpctypes.FindTxAttributes(res.TxResult.Events, hexTx)
-	if msgIndex < 0 {
+	if res.TxResult.Code != 0 {
+		// failed, return empty logs
+		return nil, nil
+	}
+
+	parsedTxs, err := rpctypes.ParseTxResult(&res.TxResult)
+	if err != nil {
+		return nil, fmt.Errorf("failed to parse tx events: %s, %v", hexTx, err)
+	}
+
+	parsedTx := parsedTxs.GetTxByHash(txHash)
+	if parsedTx == nil {
 		return nil, fmt.Errorf("ethereum tx not found in msgs: %s", hexTx)
 	}
+
 	// parse tx logs from events
-	return backend.TxLogsFromEvents(res.TxResult.Events, msgIndex)
+	return parsedTx.ParseTxLogs()
 }
 
 // Sign signs the provided data using the private key of address via Geth's signature standard.
@@ -735,15 +746,20 @@ func (e *PublicAPI) getTransactionByBlockAndIndex(block *tmrpctypes.ResultBlock,
 			e.logger.Debug("invalid ethereum tx", "height", block.Block.Header, "index", idx)
 			return nil, nil
 		}
-		// find msg index in events
-		msgIndex := rpctypes.FindTxAttributesByIndex(res.TxResult.Events, uint64(idx))
-		if msgIndex < 0 {
-			e.logger.Debug("invalid ethereum tx", "height", block.Block.Header, "index", idx)
-			return nil, nil
+
+		parsedTxs, err := rpctypes.ParseTxResult(&res.TxResult)
+		if err != nil {
+			return nil, fmt.Errorf("failed to parse tx events: %d, %v", idx, err)
 		}
+
+		parsedTx := parsedTxs.GetTxByTxIndex(int(idx))
+		if parsedTx == nil {
+			return nil, fmt.Errorf("ethereum tx not found in msgs: %d", idx)
+		}
+
 		var ok bool
 		// msgIndex is inferred from tx events, should be within bound.
-		msg, ok = tx.GetMsgs()[msgIndex].(*evmtypes.MsgEthereumTx)
+		msg, ok = tx.GetMsgs()[parsedTx.MsgIndex].(*evmtypes.MsgEthereumTx)
 		if !ok {
 			e.logger.Debug("invalid ethereum tx", "height", block.Block.Header, "index", idx)
 			return nil, nil
@@ -825,8 +841,18 @@ func (e *PublicAPI) GetTransactionReceipt(hash common.Hash) (map[string]interfac
 		return nil, nil
 	}
 
-	msgIndex, attrs := rpctypes.FindTxAttributes(res.TxResult.Events, hexTx)
-	if msgIndex < 0 {
+	// don't ignore the txs which exceed block gas limit.
+	if !backend.TxSuccessOrExceedsBlockGasLimit(&res.TxResult) {
+		return nil, nil
+	}
+
+	parsedTxs, err := rpctypes.ParseTxResult(&res.TxResult)
+	if err != nil {
+		return nil, fmt.Errorf("failed to parse tx events: %s, %v", hexTx, err)
+	}
+
+	parsedTx := parsedTxs.GetTxByHash(hash)
+	if parsedTx == nil {
 		return nil, fmt.Errorf("ethereum tx not found in msgs: %s", hexTx)
 	}
 
@@ -842,13 +868,17 @@ func (e *PublicAPI) GetTransactionReceipt(hash common.Hash) (map[string]interfac
 		return nil, fmt.Errorf("failed to decode tx: %w", err)
 	}
 
-	// the `msgIndex` is inferred from tx events, should be within the bound.
-	msg := tx.GetMsgs()[msgIndex]
-	ethMsg, ok := msg.(*evmtypes.MsgEthereumTx)
-	if !ok {
-		e.logger.Debug(fmt.Sprintf("invalid tx type: %T", msg))
-		return nil, fmt.Errorf("invalid tx type: %T", msg)
+	if res.TxResult.Code != 0 {
+		// tx failed, we should return gas limit as gas used, because that's how the fee get deducted.
+		for i := 0; i <= parsedTx.MsgIndex; i++ {
+			gasLimit := tx.GetMsgs()[i].(*evmtypes.MsgEthereumTx).GetGas()
+			parsedTxs.Txs[i].GasUsed = gasLimit
+		}
 	}
+
+	// the `msgIndex` is inferred from tx events, should be within the bound,
+	// and the tx is found by eth tx hash, so the msg type must be correct.
+	ethMsg := tx.GetMsgs()[parsedTx.MsgIndex].(*evmtypes.MsgEthereumTx)
 
 	txData, err := evmtypes.UnpackTxData(ethMsg.Data)
 	if err != nil {
@@ -862,27 +892,14 @@ func (e *PublicAPI) GetTransactionReceipt(hash common.Hash) (map[string]interfac
 		e.logger.Debug("failed to retrieve block results", "height", res.Height, "error", err.Error())
 		return nil, nil
 	}
-
 	for i := 0; i < int(res.Index) && i < len(blockRes.TxsResults); i++ {
 		cumulativeGasUsed += uint64(blockRes.TxsResults[i].GasUsed)
 	}
-	cumulativeGasUsed += rpctypes.AccumulativeGasUsedOfMsg(res.TxResult.Events, msgIndex)
-
-	var gasUsed uint64
-	if len(tx.GetMsgs()) == 1 {
-		// backward compatibility
-		gasUsed = uint64(res.TxResult.GasUsed)
-	} else {
-		gasUsed, err = rpctypes.GetUint64Attribute(attrs, evmtypes.AttributeKeyTxGasUsed)
-		if err != nil {
-			return nil, err
-		}
-	}
+	cumulativeGasUsed += parsedTxs.AccumulativeGasUsed(parsedTx.MsgIndex)
 
 	// Get the transaction result from the log
-	_, found := attrs[evmtypes.AttributeKeyEthereumTxFailed]
 	var status hexutil.Uint
-	if found {
+	if res.TxResult.Code != 0 || parsedTx.Failed {
 		status = hexutil.Uint(ethtypes.ReceiptStatusFailed)
 	} else {
 		status = hexutil.Uint(ethtypes.ReceiptStatusSuccessful)
@@ -894,28 +911,23 @@ func (e *PublicAPI) GetTransactionReceipt(hash common.Hash) (map[string]interfac
 	}
 
 	// parse tx logs from events
-	logs, err := backend.TxLogsFromEvents(res.TxResult.Events, msgIndex)
+	logs, err := parsedTx.ParseTxLogs()
 	if err != nil {
-		e.logger.Debug("logs not found", "hash", hexTx, "error", err.Error())
+		e.logger.Debug("failed to parse logs", "hash", hexTx, "error", err.Error())
 	}
 
-	// Try to find txIndex from events
-	found = false
-	txIndex, err := rpctypes.GetUint64Attribute(attrs, evmtypes.AttributeKeyTxIndex)
-	if err == nil {
-		found = true
-	} else {
+	if parsedTx.EthTxIndex == -1 {
 		// Fallback to find tx index by iterating all valid eth transactions
 		msgs := e.backend.GetEthereumMsgsFromTendermintBlock(resBlock, blockRes)
 		for i := range msgs {
 			if msgs[i].Hash == hexTx {
-				txIndex = uint64(i)
-				found = true
+				parsedTx.EthTxIndex = int64(i)
 				break
 			}
 		}
 	}
-	if !found {
+
+	if parsedTx.EthTxIndex == -1 {
 		return nil, errors.New("can't find index of ethereum tx")
 	}
 
@@ -930,14 +942,14 @@ func (e *PublicAPI) GetTransactionReceipt(hash common.Hash) (map[string]interfac
 		// They are stored in the chain database.
 		"transactionHash": hash,
 		"contractAddress": nil,
-		"gasUsed":         hexutil.Uint64(gasUsed),
+		"gasUsed":         hexutil.Uint64(parsedTx.GasUsed),
 		"type":            hexutil.Uint(txData.TxType()),
 
 		// Inclusion information: These fields provide information about the inclusion of the
 		// transaction corresponding to this receipt.
 		"blockHash":        common.BytesToHash(resBlock.Block.Header.Hash()).Hex(),
 		"blockNumber":      hexutil.Uint64(res.Height),
-		"transactionIndex": hexutil.Uint64(txIndex),
+		"transactionIndex": hexutil.Uint64(parsedTx.EthTxIndex),
 
 		// sender and receiver (contract or EOA) addreses
 		"from": from,